pyproject.toml
[project]
name = "face-generation"
version = "0.1.0"
description = "Generate faces and remove beards with GenAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.0",
    "diffusers>=0.32.1",
    "huggingface-hub>=0.27.1",
    "jupyter>=1.1.1",
    "lpips>=0.1.4",
    "matplotlib>=3.10.0",
    "mediapipe>=0.10.20",
    "numpy>=1.26.4",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "scikit-image>=0.25.0",
    "scipy>=1.15.1",
    "segmentation_models_pytorch>=0.4.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "transformers>=4.48.0",
]
